WIP - Test sketch and adapting Phoenix code
This is a WIP - I am experimenting adapting the Arduino Phoenix code
base to work with the Orion Robotics Hexapods.  This includes the Fire
Ant, which has the same type of capabilities as the Lynxmotion APod.  So
I am also merging in the changes that Zenta did to the BAP code for the
APOD.