Libraries modified/Derived for Teensy 3.1
Thought I should upload somewhere some libraries that are being modified
to support Basic Micro/Orion Robotics hardware on Teensy 3.1 based
processors.

In the case of RoboClawStream, this is an Alternate version, that is not
derived from BMSerial, but instead from Stream*.  This allows me to
simply pass in which hardware Serial port to use.  Also there were a few
casting problems fixed in this version, that fixed compiler error and
faults.

With BMSerial - This updates the library to handle the Teensy like it
does for the Arduino Due.  Only with hardware ports.  Also I think I
fixed an issue with Arduino Leonardo and other Atmega32u4 based boards,
where IO pins 0.1 are not Serial, but are instead Serial1