Make it possible to use BM PS2 on a Teensy
Previous update to BMSerial, was setup to use HardwareSerial ports if TX
and RX pins were on the right pins.  But there is no bitbang version for
the Teensy nor the Due (yet).  So updated BMSerial to allow for Half
duplex mode on Hardware Serial ports of Teensy, which support this.
Note it looks like you can do something similar to this on the Due, but
I have not done so).  Then updated BMPS2 to work on Teensy.  There was
casting issues as well as issues of having the Serial port startup at
c++ object creation time.  So I postponed doing the begin until the
first call to read the ps2